backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.4
Django==5.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-oauth-toolkit==3.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.0.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kombu==5.6.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
